Move functionality into NCDirLoader
