+++ external/fv3fit/fv3fit/data/netcdf/load.py
@@ -132,6 +132,10 @@ class NCDirLoader(TFDatasetLoader):
         nfiles: the number of files to load
         shuffle: Whether the files are opened in a shuffled order
         seed: The random seed used for filename shuffling.
+        dim_order: Order of dimensions in tensors. Will expand missing
+            dims into size 1 dim if not in a data array.
+        varying_first_dim: If true, allow number of samples to be different
+            across batches.
 
     """
 
@@ -139,6 +143,8 @@ class NCDirLoader(TFDatasetLoader):
     nfiles: Optional[int] = None
     shuffle: bool = True
     seed: int = 0
+    dim_order: Optional[Sequence[str]] = None
+    varying_first_dim: bool = False
 
     @property
     def dtype(self):
@@ -147,7 +153,17 @@ class NCDirLoader(TFDatasetLoader):
     def convert(
         self, ds: xr.Dataset, variables: Sequence[str]
     ) -> Mapping[str, tf.Tensor]:
-        return to_tensor(ds, variables, dtype=self.dtype)
+        if self.dim_order:
+            # Operates on individual data arrays to allow for inserting z dim
+            # of size 1 for 2D variables. This gives all tensors the same number
+            # of dimensions for easier concatenation later on in training.
+            tensors = {}
+            for key in variables:
+                data_array = self._ensure_consistent_dims(ds[key])
+                tensors[key] = tf.convert_to_tensor(data_array, dtype=self.dtype)
+            return tensors
+        else:
+            return to_tensor(ds, variables, dtype=self.dtype)
 
     def open_tfdataset(
         self, local_download_path: Optional[str], variable_names: Sequence[str],
@@ -162,8 +178,25 @@ class NCDirLoader(TFDatasetLoader):
             shuffle=self.shuffle,
             random_state=np.random.RandomState(self.seed),
             cache=local_download_path,
+            varying_first_dim=self.varying_first_dim,
         )
 
     @classmethod
     def from_dict(cls, d: dict) -> "TFDatasetLoader":
         return dacite.from_dict(cls, d, config=dacite.Config(strict=True))
+
+    def _ensure_consistent_dims(self, data_array: xr.DataArray):
+        if self.dim_order:
+            extra_dims_in_data_array = set(data_array.dims) - set(self.dim_order)
+            missing_dims_in_data_array = set(self.dim_order) - set(data_array.dims)
+            if len(extra_dims_in_data_array) > 0:
+                raise ValueError(
+                    f"Extra dimensions {extra_dims_in_data_array} in data that are not "
+                    f"included in configured dimension order {self.dim_order}."
+                    "Make sure these are included in the configuration dim_order."
+                )
+            da = data_array
+            for missing_dim in missing_dims_in_data_array:
+                da = data_array.expand_dims(dim=missing_dim)
+            da = da.transpose(*self.dim_order)
+        return da
